Navigation updates

+++ app/src/main/java/com/tripper/ui/homepage/list/TripListAdapter.java
@@ -115,7 +115,7 @@ public class TripListAdapter extends RecyclerView.Adapter<TripListAdapter.TripLi
                 txtEnd.setText(endDate);
 
                 long dateDiff = trip.endDate.getTimeInMillis() - trip.startDate.getTimeInMillis();
-                int dayCount = (int) dateDiff / (24 * 60 * 60 * 1000) + 1;
+                long dayCount = dateDiff / (24 * 60 * 60 * 1000) + 1;
                 String duration = String.format("Duration: %d days", dayCount);
                 txtDuration.setText(duration);
